:truck: (interface): Rename file interface/CoreLCD.hpp to interface/LCD.hpp
