+++ drivers/CoreVideo/include/CoreVideo.hpp
@@ -8,7 +8,7 @@
 #include "LKCoreSTM32HalBase.h"
 #include "interface/CoreGraphics.hpp"
 #include "interface/CoreJPEG.hpp"
-#include "interface/CoreLCD.hpp"
+#include "interface/LCD.hpp"
 #include "interface/CoreSDRAM.hpp"
 #include "interface/DMA2D.hpp"
 #include "interface/DSI.hpp"
